+++ Common/src/main/java/net/potionstudios/biomeswevegone/world/level/block/plants/bush/HydrangeaHedgeBlock.java
@@ -4,14 +4,17 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
+import net.minecraft.stats.Stats;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.gameevent.GameEvent;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.shapes.Shapes;
 import net.potionstudios.biomeswevegone.tags.BWGBlockTags;
@@ -28,10 +31,13 @@ public class HydrangeaHedgeBlock extends BWGPlacementBushBlock{
 
     @Override
     public @NotNull InteractionResult use(@NotNull BlockState state, @NotNull Level level, @NotNull BlockPos pos, Player player, @NotNull InteractionHand hand, @NotNull BlockHitResult hit) {
-        if (player.getItemInHand(hand).is(BWGItemTags.SHEARS)) {
+        ItemStack itemStack = player.getItemInHand(hand);
+        if (itemStack.is(BWGItemTags.SHEARS)) {
             level.setBlockAndUpdate(pos, BWGBlocks.HYDRANGEA_BUSH.getBlockState());
             level.playSound(null, pos, SoundEvents.SWEET_BERRY_BUSH_PICK_BERRIES, SoundSource.BLOCKS, 1.0F, 0.8F + level.random.nextFloat() * 0.4F);
             level.addParticle(ParticleTypes.HAPPY_VILLAGER, pos.getX() + level.random.nextDouble(), pos.getY() + 1.0D, pos.getZ() + level.random.nextDouble(), 0.0D, 0.0D, 0.0D);
+            level.gameEvent(player, GameEvent.SHEAR, pos);
+            player.awardStat(Stats.ITEM_USED.get(itemStack.getItem()));
             return InteractionResult.SUCCESS;
         }
         return InteractionResult.PASS;